requirements.txt
requests
azure-storage-blob==12.13.1
apache-airflow